code voor elk onderdeel apart te testen

+++ werkende_code_arduino_copy_20230428175715/werkende_code_arduino_copy_20230428175715.ino
@@ -11,7 +11,7 @@ String msg = "";
 float relatieveSnelheidH = 75;
 int branden = 0;
 float maxSnelheidV = 150; //zelf te bepalen na testen [in rpm]
-float relatieveSnelheidV = 100; //zelf te bepalen na testen
+int relatieveSnelheidV = 100; //zelf te bepalen na testen
 float msPerGraadV = (1000*255)/(maxSnelheidV*6*relatieveSnelheidV);
 
 byte sensorInterrupt = 9;  // 0 = digital pin 2
